refactor: 거래금액 타입 변경

+++ BackEnd/src/main/java/com/orange/fintech/payment/entity/TransactionMember.java
@@ -14,7 +14,7 @@ public class TransactionMember {
 
     @EmbeddedId private TransactionMemberPK transactionMemberPK;
 
-    @NotNull private Integer totalAmount;
+    @NotNull private Long totalAmount;
 
     @NotNull private Boolean isLock;
 }
